+++ tests/integration/test_zeek_dataset/test_zeek_dataset.py
@@ -5,6 +5,7 @@ from tests.common_test_utils import (
     is_evidence_present,
     create_output_dir,
     assert_no_errors,
+    close_test_redis_server,
 )
 from tests.module_factory import ModuleFactory
 import pytest
@@ -118,7 +119,7 @@ alerts_file = "alerts.log"
 
 
 @pytest.mark.parametrize(
-    "conn_log_path, expected_profiles, expected_evidence,  output_dir, redis_port",
+    "conn_log_path, expected_profiles, expected_evidence,  output_dir",
     [
         (
             "dataset/test9-mixed-zeek-dir/conn.log",
@@ -127,14 +128,12 @@ alerts_file = "alerts.log"
             "destination IP: 194.132.197.198",  # the flow with uid
             # CAwUdr34dVnyOwbUuj should trigger this
             "test9-conn_log_only/",
-            6659,
         ),
         (
             "dataset/test10-mixed-zeek-dir/conn.log",
             5,
             "non-SSL established connection",
             "test10-conn_log_only/",
-            6658,
         ),
     ],
 )
@@ -143,25 +142,31 @@ def test_zeek_conn_log(
     expected_profiles,
     expected_evidence,
     output_dir,
-    redis_port,
+    integration_port_factory,
 ):
+    redis_port = integration_port_factory("redis")
     output_dir = create_output_dir(output_dir)
+    success = False
+    try:
+        output_file = os.path.join(output_dir, "slips_output.txt")
+        command = (
+            f"./slips.py -e 1 -t -f {conn_log_path} -o {output_dir} "
+            f"-P {redis_port} > {output_file} 2>&1"
+        )
+        # this function returns when slips is done
+        run_slips(command)
+        assert_no_errors(output_dir)
 
-    output_file = os.path.join(output_dir, "slips_output.txt")
-    command = (
-        f"./slips.py -e 1 -t -f {conn_log_path} -o {output_dir} "
-        f"-P {redis_port} > {output_file} 2>&1"
-    )
-    # this function returns when slips is done
-    run_slips(command)
-    assert_no_errors(output_dir)
+        database = ModuleFactory().create_db_manager_obj(
+            redis_port, output_dir=output_dir
+        )
+        profiles = database.get_profiles_len()
+        assert profiles > expected_profiles
 
-    database = ModuleFactory().create_db_manager_obj(
-        redis_port, output_dir=output_dir
-    )
-    profiles = database.get_profiles_len()
-    assert profiles > expected_profiles
-
-    log_file = output_dir / "alerts" / alerts_file
-    assert is_evidence_present(log_file, expected_evidence) is True
-    shutil.rmtree(output_dir)
+        log_file = output_dir / "alerts" / alerts_file
+        assert is_evidence_present(log_file, expected_evidence) is True
+        success = True
+    finally:
+        if success:
+            close_test_redis_server(redis_port)
+            shutil.rmtree(output_dir)
